+++ BuildResidentialHPXML/measure.rb
@@ -295,7 +295,7 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     arg = OpenStudio::Measure::OSArgument::makeDoubleArgument('geometry_garage_protrusion', true)
     arg.setDisplayName('Geometry: Garage Protrusion')
     arg.setUnits('Frac')
-    arg.setDescription("The fraction of the garage that is protruding from the living space. Only applies to #{HPXML::ResidentialTypeSFD} units.")
+    arg.setDescription("The fraction of the garage that is protruding from the conditioned space. Only applies to #{HPXML::ResidentialTypeSFD} units.")
     arg.setDefaultValue(0.0)
     args << arg
 
@@ -1430,7 +1430,7 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     duct_leakage_units_choices << HPXML::UnitsPercent
 
     duct_location_choices = OpenStudio::StringVector.new
-    duct_location_choices << HPXML::LocationLivingSpace
+    duct_location_choices << HPXML::LocationConditionedSpace
     duct_location_choices << HPXML::LocationBasementConditioned
     duct_location_choices << HPXML::LocationBasementUnconditioned
     duct_location_choices << HPXML::LocationCrawlspace
@@ -1796,7 +1796,7 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     water_heater_fuel_choices << HPXML::FuelTypeCoal
 
     water_heater_location_choices = OpenStudio::StringVector.new
-    water_heater_location_choices << HPXML::LocationLivingSpace
+    water_heater_location_choices << HPXML::LocationConditionedSpace
     water_heater_location_choices << HPXML::LocationBasementConditioned
     water_heater_location_choices << HPXML::LocationBasementUnconditioned
     water_heater_location_choices << HPXML::LocationGarage
@@ -2216,7 +2216,7 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     args << arg
 
     battery_location_choices = OpenStudio::StringVector.new
-    battery_location_choices << HPXML::LocationLivingSpace
+    battery_location_choices << HPXML::LocationConditionedSpace
     battery_location_choices << HPXML::LocationBasementConditioned
     battery_location_choices << HPXML::LocationBasementUnconditioned
     battery_location_choices << HPXML::LocationCrawlspace
@@ -2406,7 +2406,7 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     args << arg
 
     appliance_location_choices = OpenStudio::StringVector.new
-    appliance_location_choices << HPXML::LocationLivingSpace
+    appliance_location_choices << HPXML::LocationConditionedSpace
     appliance_location_choices << HPXML::LocationBasementConditioned
     appliance_location_choices << HPXML::LocationBasementUnconditioned
     appliance_location_choices << HPXML::LocationGarage
@@ -4083,7 +4083,7 @@ class HPXMLFile
         if adjacent_surface.nil? # adjacent to a space that is not explicitly in the model
           unless [HPXML::ResidentialTypeSFD].include?(args[:geometry_unit_type])
             exterior_adjacent_to = interior_adjacent_to
-            if exterior_adjacent_to == HPXML::LocationLivingSpace # living adjacent to living
+            if exterior_adjacent_to == HPXML::LocationConditionedSpace # conditioned space adjacent to conditioned space
               exterior_adjacent_to = HPXML::LocationOtherHousingUnit
             end
           end
@@ -4126,16 +4126,16 @@ class HPXMLFile
       next if Geometry.surface_is_rim_joist(surface, args[:geometry_rim_joist_height])
 
       interior_adjacent_to = Geometry.get_adjacent_to(surface: surface)
-      next unless [HPXML::LocationLivingSpace, HPXML::LocationAtticUnvented, HPXML::LocationAtticVented, HPXML::LocationGarage].include? interior_adjacent_to
+      next unless [HPXML::LocationConditionedSpace, HPXML::LocationAtticUnvented, HPXML::LocationAtticVented, HPXML::LocationGarage].include? interior_adjacent_to
 
       exterior_adjacent_to = HPXML::LocationOutside
       if surface.adjacentSurface.is_initialized
         exterior_adjacent_to = Geometry.get_adjacent_to(surface: surface.adjacentSurface.get)
-      elsif surface.outsideBoundaryCondition == 'Adiabatic' # can be adjacent to living space, attic
+      elsif surface.outsideBoundaryCondition == 'Adiabatic' # can be adjacent to conditioned space, attic
         adjacent_surface = Geometry.get_adiabatic_adjacent_surface(model: model, surface: surface)
         if adjacent_surface.nil? # adjacent to a space that is not explicitly in the model
           exterior_adjacent_to = interior_adjacent_to
-          if exterior_adjacent_to == HPXML::LocationLivingSpace # living adjacent to living
+          if exterior_adjacent_to == HPXML::LocationConditionedSpace # conditioned space adjacent to conditioned space
             exterior_adjacent_to = HPXML::LocationOtherHousingUnit
           end
         else # adjacent to a space that is explicitly in the model
@@ -4143,7 +4143,7 @@ class HPXMLFile
         end
       end
 
-      next if exterior_adjacent_to == HPXML::LocationLivingSpace # already captured these surfaces
+      next if exterior_adjacent_to == HPXML::LocationConditionedSpace # already captured these surfaces
 
       attic_locations = [HPXML::LocationAtticUnconditioned, HPXML::LocationAtticUnvented, HPXML::LocationAtticVented]
       attic_wall_type = nil
@@ -4217,7 +4217,7 @@ class HPXMLFile
         if adjacent_surface.nil? # adjacent to a space that is not explicitly in the model
           unless [HPXML::ResidentialTypeSFD].include?(args[:geometry_unit_type])
             exterior_adjacent_to = interior_adjacent_to
-            if exterior_adjacent_to == HPXML::LocationLivingSpace # living adjacent to living
+            if exterior_adjacent_to == HPXML::LocationConditionedSpace # conditioned space adjacent to conditioned space
               exterior_adjacent_to = HPXML::LocationOtherHousingUnit
             end
           end
@@ -4306,7 +4306,7 @@ class HPXMLFile
       next unless ['Floor', 'RoofCeiling'].include? surface.surfaceType
 
       interior_adjacent_to = Geometry.get_adjacent_to(surface: surface)
-      next unless [HPXML::LocationLivingSpace, HPXML::LocationGarage].include? interior_adjacent_to
+      next unless [HPXML::LocationConditionedSpace, HPXML::LocationGarage].include? interior_adjacent_to
 
       exterior_adjacent_to = HPXML::LocationOutside
       if surface.adjacentSurface.is_initialized
@@ -4322,7 +4322,7 @@ class HPXMLFile
 
       next if interior_adjacent_to == exterior_adjacent_to
       next if (surface.surfaceType == 'RoofCeiling') && (exterior_adjacent_to == HPXML::LocationOutside)
-      next if [HPXML::LocationLivingSpace,
+      next if [HPXML::LocationConditionedSpace,
                HPXML::LocationBasementConditioned,
                HPXML::LocationCrawlspaceConditioned].include? exterior_adjacent_to
 
@@ -4579,10 +4579,10 @@ class HPXMLFile
     end
 
     # Add attached roofs for cathedral ceiling
-    living_space = HPXML::LocationLivingSpace
+    conditioned_space = HPXML::LocationConditionedSpace
     surf_ids['roofs']['surfaces'].each do |surface|
-      next if (living_space != surface.interior_adjacent_to) &&
-              (living_space != surface.exterior_adjacent_to)
+      next if (conditioned_space != surface.interior_adjacent_to) &&
+              (conditioned_space != surface.exterior_adjacent_to)
 
       surf_ids['roofs']['ids'] << surface.id
     end
@@ -4611,7 +4611,7 @@ class HPXMLFile
       surf_hash['surfaces'].each do |surface|
         next unless (foundation_locations.include? surface.interior_adjacent_to) ||
                     (foundation_locations.include? surface.exterior_adjacent_to) ||
-                    (surf_type == 'slabs' && surface.interior_adjacent_to == HPXML::LocationLivingSpace) ||
+                    (surf_type == 'slabs' && surface.interior_adjacent_to == HPXML::LocationConditionedSpace) ||
                     (surf_type == 'floors' && [HPXML::LocationOutside, HPXML::LocationManufacturedHomeUnderBelly].include?(surface.exterior_adjacent_to))
 
         surf_hash['ids'] << surface.id
@@ -5094,7 +5094,7 @@ class HPXMLFile
       elsif attic_type == HPXML::AtticTypeVented
         return HPXML::LocationAtticVented
       elsif attic_type == HPXML::AtticTypeConditioned
-        return HPXML::LocationLivingSpace
+        return HPXML::LocationConditionedSpace
       else
         fail "Specified '#{location}' but attic type is '#{attic_type}'."
       end
@@ -5129,7 +5129,7 @@ class HPXMLFile
 
     if (not ducts_supply_location.nil?) && ducts_supply_surface_area.nil? && ducts_supply_area_fraction.nil?
       # Supply duct location without any area inputs provided; set area fraction
-      if ducts_supply_location == HPXML::LocationLivingSpace
+      if ducts_supply_location == HPXML::LocationConditionedSpace
         ducts_supply_area_fraction = 1.0
       else
         ducts_supply_area_fraction = HVAC.get_default_duct_fraction_outside_conditioned_space(args[:geometry_unit_num_floors_above_grade])
@@ -5138,7 +5138,7 @@ class HPXMLFile
 
     if (not ducts_return_location.nil?) && ducts_return_surface_area.nil? && ducts_return_area_fraction.nil?
       # Return duct location without any area inputs provided; set area fraction
-      if ducts_return_location == HPXML::LocationLivingSpace
+      if ducts_return_location == HPXML::LocationConditionedSpace
         ducts_return_area_fraction = 1.0
       else
         ducts_return_area_fraction = HVAC.get_default_duct_fraction_outside_conditioned_space(args[:geometry_unit_num_floors_above_grade])
@@ -5172,21 +5172,21 @@ class HPXMLFile
     end
 
     if (not ducts_supply_area_fraction.nil?) && (ducts_supply_area_fraction < 1)
-      # OS-HPXML needs duct fractions to sum to 1; add remaining ducts in living space.
+      # OS-HPXML needs duct fractions to sum to 1; add remaining ducts in conditioned space.
       hvac_distribution.ducts.add(id: "Ducts#{hvac_distribution.ducts.size + 1}",
                                   duct_type: HPXML::DuctTypeSupply,
                                   duct_insulation_r_value: 0.0,
-                                  duct_location: HPXML::LocationLivingSpace,
+                                  duct_location: HPXML::LocationConditionedSpace,
                                   duct_fraction_area: 1.0 - ducts_supply_area_fraction)
     end
 
     if not hvac_distribution.ducts.find { |d| d.duct_type == HPXML::DuctTypeReturn }.nil?
       if (not ducts_return_area_fraction.nil?) && (ducts_return_area_fraction < 1)
-        # OS-HPXML needs duct fractions to sum to 1; add remaining ducts in living space.
+        # OS-HPXML needs duct fractions to sum to 1; add remaining ducts in conditioned space.
         hvac_distribution.ducts.add(id: "Ducts#{hvac_distribution.ducts.size + 1}",
                                     duct_type: HPXML::DuctTypeReturn,
                                     duct_insulation_r_value: 0.0,
-                                    duct_location: HPXML::LocationLivingSpace,
+                                    duct_location: HPXML::LocationConditionedSpace,
                                     duct_fraction_area: 1.0 - ducts_return_area_fraction)
       end
     end
@@ -5912,7 +5912,7 @@ class HPXMLFile
                             integrated_energy_factor: integrated_energy_factor,
                             rh_setpoint: args[:dehumidifier_rh_setpoint],
                             fraction_served: args[:dehumidifier_fraction_dehumidification_load_served],
-                            location: HPXML::LocationLivingSpace)
+                            location: HPXML::LocationConditionedSpace)
   end
 
   def self.set_clothes_washer(hpxml, args)